fix for ArrayIndexOutOfBoundsException as records size exceeds 128

+++ tetrapak-customerhub/core/src/main/java/com/tetrapak/customerhub/core/services/impl/FinancialResultsExcelServiceImpl.java
@@ -234,7 +234,7 @@ public class FinancialResultsExcelServiceImpl implements FinancialResultsExcelSe
         if (null != records && !records.isEmpty()) {
             data = new String[records.size()][16];
             Iterator<Record> itr = records.iterator();
-            byte counter = 0;
+            int counter = 0;
             String[] tags = new String[]{};
             while (itr.hasNext()) {
                 Record record = itr.next();
